EOEPCA-179 Initial Structure of the SDD/ICD docs

+++ docs/bin/publish-docs.sh
@@ -4,17 +4,10 @@ ORIG_DIR="$(pwd)"
 cd "$(dirname "$0")"
 BIN_DIR="$(pwd)"
 
-trap "cd '${ORIG_DIR}'" EXIT
+trap "cd '${ORIG_DIR}';" EXIT
 
 # Work in the docs/ directory
 cd "${BIN_DIR}/.."
-
-# Check the output directory exists
-if [ ! -d "output" ]
-then
-  echo "ERROR: output directory is missing, generate-docs must be run first" >&2
-  exit 1
-fi
 
 # Deduce the name of the repository
 if [ -z "${GH_REPOS_NAME}" ]
@@ -25,16 +18,26 @@ fi
 # Clone the 'gh-pages' branch
 git clone --branch gh-pages --single-branch "https://${GH_TOKEN}@github.com/EOEPCA/${GH_REPOS_NAME}" repos
 
-# Move generated doc outputs to the repos
+for doc in SDD ICD; do
+	# Check the output directory exists
+	if [ ! -d "$doc/output" ]
+		then
+		  echo "ERROR: output directory is missing, generate-docs must be run first" >&2
+		exit 1
+	fi
+	cd repos/$doc
+	rm -rf current
+	mv ../../$doc/output current
+	mv current/index.pdf current/EOEPCA-${GH_REPOS_NAME}.pdf
+
+	# Prepare the root landing page/README - but don't overwrite if they already exist
+	#if [ ! -e index.html ]; then cp ../gh-page-root.html index.html; fi
+	#if [ ! -e README.adoc ]; then cp ../gh-page-README.adoc README.adoc; fi
+	cd ${BIN_DIR}/..
+
+done
+
 cd repos
-rm -rf current
-mv ../output current
-mv current/index.pdf current/EOEPCA-${GH_REPOS_NAME}.pdf
-
-# Prepare the root landing page/README - but don't overwrite if they already exist
-if [ ! -e index.html ]; then cp ../gh-page-root.html index.html; fi
-if [ ! -e README.adoc ]; then cp ../gh-page-README.adoc README.adoc; fi
-
 # Config git profile for commits
 if [ -n "${GH_USER_NAME}" ]; then git config user.name "${GH_USER_NAME}"; fi
 if [ -n "${GH_USER_EMAIL}" ]; then git config user.email "${GH_USER_EMAIL}"; fi
